# api/docs/swagger.yaml
basePath: /
definitions:
...
  db.MatchType:
    enum:
    - SINGLES
    - DOUBLES
    - MIXED
    type: string
    x-enum-varnames:
    - MatchTypeSINGLES
    - MatchTypeDOUBLES
    - MatchTypeMIXED
  db.User:
    properties:
      created_at:
        $ref: '#/definitions/pgtype.Timestamptz'
      elo_doubles:
        type: integer
      elo_mixed:
        type: integer
      elo_singles:
        type: integer
      email:
        type: string
      ffbad_rank:
        $ref: '#/definitions/pgtype.Text'
      first_name:
        type: string
      gender:
        $ref: '#/definitions/pgtype.Text'
      id:
        type: string
      last_name:
        type: string
      role:
        type: string
    type: object
...
  handlers.updateUserRequest:
    properties:
      email:
        type: string
      ffbad_rank:
        description: '"" | NC | P12 | ... | N1'
        type: string
      first_name:
        type: string
      gender:
        description: '"" | MALE | FEMALE'
        type: string
      last_name:
        type: string
    type: object
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    format: int32
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
...
  pgtype.Text:
    properties:
      string:
        type: string
      valid:
        type: boolean
    type: object
  pgtype.Timestamptz:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        type: string
      valid:
        type: boolean
    type: object
  services.EloRanking:
    properties:
      elo:
        type: integer
      rank:
        type: integer
      tableau:
        $ref: '#/definitions/db.MatchType'
      user:
        $ref: '#/definitions/db.User'
    type: object
  services.GiantKillerRanking:
    properties:
      rank:
        type: integer
      upset_wins:
        type: integer
      user:
        $ref: '#/definitions/db.User'
    type: object
  services.LeagueRanking:
    properties:
      losses:
        type: integer
      rank:
        type: integer
      sets_diff:
        type: integer
      sets_lost:
        type: integer
      sets_won:
        type: integer
      user:
        $ref: '#/definitions/db.User'
      wins:
        type: integer
    type: object
  services.MatchesPlayedRanking:
    properties:
      matches_played:
        type: integer
      rank:
        type: integer
      user:
        $ref: '#/definitions/db.User'
    type: object
  services.PerformanceRanking:
    properties:
      points:
        type: integer
      rank:
        type: integer
      user:
        $ref: '#/definitions/db.User'
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for the BAP Pulse badminton club app — Bad A Paname.
  title: BAP Pulse API
  version: "1.0"
paths:
...
  /rankings:
    get:
      description: Use /rankings/elo?tableau=... for new clients.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.EloRanking'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ELO leaderboard (legacy — defaults to SINGLES)
      tags:
      - rankings
  /rankings/elo:
    get:
      parameters:
      - description: SINGLES (default) | DOUBLES | MIXED
        in: query
        name: tableau
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.EloRanking'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ELO leaderboard for a tableau
      tags:
      - rankings
  /rankings/giant-killer:
    get:
      description: Wins against an opponent whose ELO at match time exceeded the player's
        by at least UpsetThreshold.
      parameters:
      - description: SINGLES (default) | DOUBLES | MIXED
        in: query
        name: tableau
        type: string
      - description: 'YYYY-MM (default: current month)'
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.GiantKillerRanking'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Giant-killer ranking
      tags:
      - rankings
  /rankings/league:
    get:
      parameters:
      - description: SINGLES (default) | DOUBLES | MIXED
        in: query
        name: tableau
        type: string
      - description: 'YYYY-MM (default: current month)'
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.LeagueRanking'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: League ranking (wins / losses / sets)
      tags:
      - rankings
  /rankings/matches-played:
    get:
      parameters:
      - description: SINGLES (default) | DOUBLES | MIXED
        in: query
        name: tableau
        type: string
      - description: 'YYYY-MM (default: current month)'
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.MatchesPlayedRanking'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Matches-played ranking
      tags:
      - rankings
  /rankings/performance:
    get:
      description: Sum of performance points awarded during the period for the tableau.
      parameters:
      - description: SINGLES (default) | DOUBLES | MIXED
        in: query
        name: tableau
        type: string
      - description: 'YYYY-MM (default: current month)'
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.PerformanceRanking'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Performance ranking (yellow jersey)
      tags:
      - rankings
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - users
  /users/{id}:
    delete:
      description: Users can only delete their own account. Admins can delete any
        account.
      parameters:
      - description: User ID (Firebase UID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete user account
      tags:
      - users
    get:
      parameters:
      - description: User ID (Firebase UID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.User'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Users can only update their own profile. Admins can update any
        profile.
      parameters:
      - description: User ID (Firebase UID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated user data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.updateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'Firebase ID Token — format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"